Consolidate comparisons funcs into one file
